+++ etl/cite/itp_cite_empresas_tipo_agg.py
@@ -85,7 +85,7 @@ class CiteEmpresasPipeline(EasyPipeline):
         dtypes = {
 
             'empresa_id':            'UInt8',
-            'time':                  'UInt16',
+            'time':                  'UInt32',
             'empresas':              'Float32',
          }
